Add files via upload
To compare the CPU scheduling algorithms such as FCFS(First Come First Serve),SJF(Shortest Job First,Round Robin on the basis of average waiting time and average turnaround time with the help of a JavaFx library. 
In this project MySql Database is used to input process and Java Fx library is used to create a gui for visualization of data.